fett/requirements.txt
asgiref
beautifulsoup4
certifi
charset-normalizer
clean-text
click
colorama
Django==3.2
djangorestframework==3.13.1
emoji
ftfy
html5lib
idna
joblib
nltk
pytz
regex
requests
six
soupsieve
sqlparse
tqdm
tzdata
unicode
urllib3
wcwidth
webencodings
django-cors-headers